llm_benchmark_agent_v1/requirements.txt
openai>=1.60.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
jinja2>=3.1.0
requests>=2.31.0
litellm>=1.55.0
rich>=13.0.0
click>=8.1.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.1.0
